Fixed major bug: new price was never saved.
During testing, I used the difference 0 (zero) to
cause a notification to be sent for each new value
instead of having to wait for it to change. I did
not see that the value wasn't actually being
updated in the bolt database until just a few
days ago.

This saves the new price in the bolt db to check
against after at each interval.

Will commit a cleaner solution soon for the price
checking logic.