Added etc directory for build. Fix email location.
The email template file email.tmpl was trying to be
loaded from the working process's directory. The
etc directory in the binary's build location stores
the config.yaml file and the email.tmpl file and
will load those files if they do not exist in the
current working directory.

Embedding the data in email.tmpl as a constant
might be more efficient and easier to distribute,
but this allows the email template to be updated
on the spot without restarting the service.